// src/types/Color.ts
export enum EnumColorVariant {
    DARK = 'dark',
    LIGHT = 'light',
    BLUE = 'blue',
    GREEN = 'green',
    YELLOW = 'yellow',
    RED = 'red',
}

export enum EnumColorHex {
    BLUE = '#4285f4',
    GREEN = '#34a853',
    YELLOW = '#f9ab00',
    RED = '#ea4335',
    BLUE_HALFTONE = '#57caff',
    GREEN_HALFTONE = '#5cdb6d',
    YELLOW_HALFTONE = '#ffd427',
    RED_HALFTONE = '#ff7daf',
    BLUE_PASTEL = '#c3ecf6',
    GREEN_PASTEL = '#ccf6c5',
    YELLOW_PASTEL = '#ffe7a5',
    RED_PASTEL = '#f8d8d8',
    GREYSCALE_WHITE = '#f0f0f0',
    GREYSCALE_BLACK = '#1e1e1e',
}